refactor(ui): extract file-modifying tools + slim message-part aggregator
Slice 11b.1 steps 13-14 of the message-part.tsx split. Final extraction:
edit / write / apply_patch renderers move to a dedicated module, and
message-part.tsx collapses to a thin aggregator + public barrel.

New: packages/ui/src/components/message-part-tools-file.tsx
  Side-effect registers: edit · write · apply_patch
  Imports ToolFileAccordion + DiagnosticsDisplay + getDiagnostics from
  ./message-part-tool-display rather than redeclaring them.

message-part.tsx now: 91 lines (was 2459 pre-slice). Owns only the role
dispatcher (Message). Public re-exports preserve the legacy import path
for session-turn.tsx, storybook, and tests. Concrete renderers register
themselves through side-effect imports.

All 14 modules under the 500-line AGENTS.md ceiling:
   53  message-part-types.ts
   57  message-part-session-link.ts
   75  message-part-registry.tsx
   85  message-part-tools-shell.tsx
   91  message-part.tsx
  102  message-part-group.ts
  104  message-part-tools-agent.tsx
  167  message-part-render-groups.ts
  171  message-part-markdown.tsx
  177  message-part-core-renderers.tsx
  197  message-part-tool-info.ts
  244  message-part-tool-display.tsx
  355  message-part-tools-file.tsx
  396  message-part-tools-basic.tsx

tsgo --noEmit GREEN. bun test (ui): 478 pass / 3 pre-existing fail.
bun test (app): 1050 pass / 0 fail.
